tweak
git-svn-id: https://svn.r-project.org/R/trunk@51956 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/tools/R/install.R
@@ -482,9 +482,11 @@
                 if (debug) starsmsg(stars, "backing up earlier installation")
                 if(WINDOWS) {
                     file.copy(instdir, lockdir, recursive = TRUE)
-                    unlink(instdir, recursive = TRUE)
-                } else
+                    if (more_than_libs) unlink(instdir, recursive = TRUE)
+                } else if (more_than_libs)
                     system(paste("mv", instdir, file.path(lockdir, pkg_name)))
+                else
+                    file.copy(instdir, lockdir, recursive = TRUE)
             } else if (more_than_libs) unlink(instdir, recursive = TRUE)
             dir.create(instdir, recursive = TRUE, showWarnings = FALSE)
         }
